crimson/os/seastore/journal: relocate RecordSubmitter related classes
